chore: refract code
