+++ lib/screens/settings/myprofile.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class MyProfile extends StatelessWidget {
+  const MyProfile({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      child: Column(
+        children: const [
+          Text("Akash"),
+          Text("Email"),
+        ],
+      ),
+    );
+  }
+}
