--- lib/screens/settings/settings.dart
@@ -1,25 +1,0 @@
-import 'package:flutter/material.dart';
-import 'components/body.dart';
-
-class SettingPage extends StatefulWidget {
-  const SettingPage({Key? key}) : super(key: key);
-  @override
-  SettingPageState createState() => SettingPageState();
-}
-
-class SettingPageState extends State<SettingPage> {
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text("Settings"),
-      ),
-      body: const Body(),
-    );
-  }
-}
